rename oldcif
